dev-requirements.txt
Django>=4.2,<5.0
djangorestframework>=3.14.0
django-cors-headers>=4.2.0
nltk>=3.8.1
newspaper3k>=0.2.8
lxml>=5.2.0
lxml[html_clean]>=5.2.0
gunicorn>=21.2.0
